requirements.txt
streamlit>=1.40,<2
gspread>=6.1,<7
google-auth>=2.35,<3
bcrypt>=4.2,<5
pandas>=2.2,<3
pytz>=2024.1,<2027
